requirements.txt
# Core Streamlit and Data Processing
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0

# Google APIs
google-ads>=22.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Environment
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0